requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
requests>=2.31.0
aiohttp>=3.9.0
millify>=0.1.1
nltk>=3.8.1
python-dotenv>=1.0.0
pyarrow>=14.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
openai>=1.12.0
exa-py>=1.0.0
duckduckgo-search>=4.1.0
agno>=2.2.7
fmp-python>=0.1.5
fastmcp>=2.13.0
yfinance>=0.2.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0

